separate code index.js and utils.js
